Fix build errors in Native MSMQ implementation
- Fixed VT_VECTOR constant: byte -> ushort (4096 > 255)
- Fixed EventLog platform warning with OperatingSystem.IsWindows()
- Both apps now build cleanly on Windows
- Ready for deployment

+++ ReceiverWebApp/Services/NativeMsmq.cs
@@ -24,8 +24,8 @@ public static class NativeMsmq
     private const int PROPID_M_BODY_TYPE = 42;
     private const int PROPID_M_LABEL = 11;
     
-    private const byte VT_VECTOR = 0x1000;
-    private const byte VT_UI1 = 17;
+    private const ushort VT_VECTOR = 0x1000;
+    private const ushort VT_UI1 = 17;
     
     [StructLayout(LayoutKind.Sequential)]
     private struct MQMSGPROPS
